Add tests to verify that cache creation/invalidation happens when cached scripts change in a variety of ways
