+++ amm/src/test/resources/scripts/caching/scalatagsPredef.scala
@@ -1,0 +1,1 @@
+import $ivy.`com.lihaoyi::scalatags:0.5.4`, scalatags.Text.all._
