+++ amm/src/test/resources/scripts/caching/scalatagsScript.scala
@@ -1,0 +1,1 @@
+println(div("<('.'<)").render)
